Move ppu logic into a separate file to keep lib.rs simpler.
